split the script event codebase
